requirements.txt
Django==1.6
MySQL-python==1.2.4
South==0.8.1
decorator==3.4.0
dj-database-url==0.2.2
django-appconf==0.6
django-compressor==1.3
django-crispy-forms==1.4
django-debug-toolbar==1.0.1
django-grappelli==2.4.5
django-pylibmc==0.5.0
django-settings
django-storages
boto
django-boto
django-s3-folder-storage
gunicorn==18.0
pylibmc==1.2.3
python-memcached==1.53
six==1.5.2
sqlparse==0.1.11
wsgiref==0.1.2
psycopg2
newrelic
beautifulsoup4
requests
python-dateutil